refactor(plugins): move betterAuthPlugin to src/plugins/auth/ (CP-1)
- src/lib/auth-plugin.ts -> src/plugins/auth/auth-plugin.ts
- src/lib/__tests__/feature-guard.test.ts
    -> src/plugins/auth/__tests__/feature-guard.test.ts
- src/lib/__tests__/auth-plugin-unauthorized-log.test.ts
    -> src/plugins/auth/__tests__/unauthorized-log.test.ts

Bulk-updated 40 import sites across src/modules/**/index.ts,
controllers, and test harnesses. Relative imports inside the plugin
(`./auth`, `./errors/http-errors`, `./logger`, `./permissions`)
rewritten to absolute `@/lib/...` since the plugin moved directories.

lib/auth.ts (856 lines — the Better Auth config) stays put; CP-4
will break it down into sub-files alongside the plugin.

CLAUDE.md references in api-keys and auth modules updated to the
new path.

Tests: 39 pass across plugins/auth (unauthorized-log) and api-keys
(consumer of betterAuthPlugin macro).
