+++ src/index.ts
@@ -7,7 +7,6 @@ import { rateLimit } from "elysia-rate-limit";
 import { z } from "zod";
 import { pool } from "./db";
 import { env, isProduction } from "./env";
-import { betterAuthPlugin, OpenAPI } from "./lib/auth-plugin";
 import { parseOrigins } from "./lib/cors";
 import { logger } from "./lib/logger";
 import { setupGracefulShutdown } from "./lib/shutdown/shutdown";
@@ -20,6 +19,7 @@ import { organizationController } from "./modules/organizations";
 import { paymentsController } from "./modules/payments";
 import { registerPaymentListeners } from "./modules/payments/hooks/listeners";
 import { publicController } from "./modules/public";
+import { betterAuthPlugin, OpenAPI } from "./plugins/auth/auth-plugin";
 import { cronPlugin } from "./plugins/cron/cron-plugin";
 import { errorPlugin } from "./plugins/errors/error-plugin";
 import { healthPlugin } from "./plugins/health/health-plugin";
