+++ src/modules/admin/api-keys/index.ts
@@ -1,6 +1,5 @@
 import { Elysia } from "elysia";
 import { isProduction } from "@/env";
-import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { wrapSuccess } from "@/lib/responses/envelope";
 import {
   forbiddenErrorSchema,
@@ -8,6 +7,7 @@ import {
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
+import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
 import {
   createApiKeyResponseSchema,
   createApiKeySchema,
