+++ src/modules/cbo-occupations/index.ts
@@ -1,11 +1,11 @@
 import { Elysia } from "elysia";
-import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { wrapSuccess } from "@/lib/responses/envelope";
 import {
   notFoundErrorSchema,
   unauthorizedErrorSchema,
   validationErrorSchema,
 } from "@/lib/responses/response.types";
+import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
 import {
   cboOccupationIdParamSchema,
   getCboOccupationResponseSchema,
