+++ src/modules/occurrences/cpf-analyses/index.ts
@@ -1,6 +1,5 @@
 import { Elysia } from "elysia";
 import { isProduction } from "@/env";
-import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { wrapSuccess } from "@/lib/responses/envelope";
 import {
   conflictErrorSchema,
@@ -10,6 +9,7 @@ import {
   validationErrorSchema,
 } from "@/lib/responses/response.types";
 import { auditPlugin } from "@/plugins/audit/audit-plugin";
+import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
 import {
   createCpfAnalysisResponseSchema,
   createCpfAnalysisSchema,
