+++ src/modules/payments/pagarme/pagarme-orphaned-plans.controller.ts
@@ -1,11 +1,11 @@
 import { Elysia } from "elysia";
 import { isProduction } from "@/env";
-import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { wrapSuccess } from "@/lib/responses/envelope";
 import {
   forbiddenErrorSchema,
   unauthorizedErrorSchema,
 } from "@/lib/responses/response.types";
+import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
 import {
   cleanupOrphanedPlansResponseSchema,
   listOrphanedPlansResponseSchema,
