+++ src/test/support/app.ts
@@ -1,7 +1,6 @@
 import { cors } from "@elysiajs/cors";
 import { Elysia } from "elysia";
 import { env } from "@/env";
-import { betterAuthPlugin } from "@/lib/auth-plugin";
 import { adminController } from "@/modules/admin";
 import { auditController } from "@/modules/audit";
 import { employeeController } from "@/modules/employees";
@@ -9,6 +8,7 @@ import { occurrencesController } from "@/modules/occurrences";
 import { organizationController } from "@/modules/organizations";
 import { paymentsController } from "@/modules/payments";
 import { publicController } from "@/modules/public";
+import { betterAuthPlugin } from "@/plugins/auth/auth-plugin";
 import { errorPlugin } from "@/plugins/errors/error-plugin";
 import { healthPlugin } from "@/plugins/health/health-plugin";
 import { loggerPlugin } from "@/plugins/logger/logger-plugin";
